# android/data/mobile-api-client/openapi/mobile-v1.yaml
components:
  schemas:
...
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/mobile/v1/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          example: Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          nullable: true
          type: array
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          example: https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          example: 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          example: Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          example: https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    FileReadResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/mobile/v1/schemas/FileReadResponse.json
          format: uri
          readOnly: true
          type: string
        content:
          type: string
        language:
          description: Hint de linguagem derivado da extensão, para syntax highlighting
          type: string
        mtime:
          description: mtime do arquivo em disco (unix) — devolver em write.expected_mtime para detectar conflito
          format: int64
          type: integer
        size:
          format: int64
          type: integer
      required:
        - content
        - mtime
        - language
        - size
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Access token emitido por POST /auth/login (ou renovado por POST /auth/refresh), enviado como `Authorization: Bearer <token>`. Quem valida é auth.Middleware, o mesmo de toda rota autenticada do painel."
      scheme: bearer
      type: http
info:
  title: server-control-panel mobile BFF
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /files/read:
    get:
      operationId: readFile
      parameters:
        - description: Caminho absoluto do arquivo a ler
          explode: false
          in: query
          name: path
          required: true
          schema:
            description: Caminho absoluto do arquivo a ler
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileReadResponse"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Not Found
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Request Entity Too Large
        "415":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Unsupported Media Type
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Internal Server Error
      security:
        - bearerAuth: []
      summary: Lê um arquivo de texto do servidor com hint de linguagem para highlight
      tags:
        - mobile
...
servers:
  - url: /api/mobile/v1